LPS-53578 rename navigation an navigation menu
