+++ modules/apps/site-navigation/site-navigation-menu-web/src/META-INF/resources/view.jsp
@@ -18,7 +18,7 @@
 
 <liferay-ui:navigation
 	bulletStyle="<%= bulletStyle %>"
-	displayStyleDefinition='<%= NavigationWebConfigurationUtil.getArray("display.style", new Filter(displayStyle)) %>'
+	displayStyleDefinition='<%= NavigationMenuWebConfigurationUtil.getArray("display.style", new Filter(displayStyle)) %>'
 	headerType="<%= headerType %>"
 	includedLayouts="<%= includedLayouts %>"
 	nestedChildren="<%= nestedChildren %>"
