+++ modules/apps/site-navigation/site-navigation-menu-web/src/content/Language.properties
@@ -1,4 +1,4 @@
-javax.portlet.title.com_liferay_site_navigation_menu_web_portlet_NavigationPortlet=Navigation
+javax.portlet.title.com_liferay_site_navigation_menu_web_portlet_NavigationPortlet=Navigation Menu
 bullet-style=Bullet Style
 from-level-0=From Level 0
 from-level-1=From Level 1
